Refactor: Combined PrivateRoute and LoginRoute
